refactor animation states
